\name{blocfit}
\Rdversion{1.1}
\alias{blocfit}
\title{locfit reimplemented from first principals.}
\description{
Performs locally weighted polynomial regression. 
}
\usage{
blocfit(x, y, a = 0.7, deg = 1, kern = 'bisq')
}
\arguments{
  \item{x}{}
  \item{y}{}
  \item{a}{}
  \item{deg}{}
  \item{kern}{}
}
\seealso{
  \code{\link{oneD}}.
}
\examples{
#
}
\keyword{file}
